requirements.txt
# requirements.txt for Modular RAG Framework
python-dotenv
pyyaml
pandas>=1.4,<3.0
sqlalchemy>=1.4,<3.0
pyhive[hive]
openai>=1.0.0
chromadb==0.4.22
streamlit>=1.25
numpy>=1.23,<2.0.0
psycopg2-binary
mysql-connector-python
loguru
# For MCP integration (if needed)
requests
